Made Tyrian-Tags, Tyrian module now JS only
