+++ tyrian-tags/js/src/main/scala/tyrian/Tyrian.scala
@@ -1,0 +1,8 @@
+package tyrian
+
+object Tyrian:
+
+  type Event            = org.scalajs.dom.Event
+  type KeyboardEvent    = org.scalajs.dom.KeyboardEvent
+  type MouseEvent       = org.scalajs.dom.MouseEvent
+  type HTMLInputElement = org.scalajs.dom.HTMLInputElement
