+++ tyrian-tags/js/src/main/scala/tyrian/syntax.scala
@@ -1,0 +1,5 @@
+package tyrian
+
+object syntax:
+
+  extension [M](oa: Option[Elem[M]]) def orEmpty: Elem[M] = oa.getOrElse(tyrian.Empty)
