+++ tyrian-tags/jvm/src/main/scala/tyrian/LocationPlatform.scala
@@ -1,3 +1,4 @@
 package tyrian
 
 trait LocationPlatform
+
