+++ tyrian-tags/shared/src/main/scala/tyrian/LocationDetails.scala
