+++ tyrian-tags/shared/src/test/scala/tyrian/LocationDetailsTests.scala
