+++ tyrian/src/main/scala/tyrian/Cmd.scala
