+++ tyrian/src/main/scala/tyrian/TyrianAppF.scala
@@ -6,6 +6,7 @@ import org.scalajs.dom.Element
 import org.scalajs.dom.HTMLElement
 import org.scalajs.dom.document
 import org.scalajs.dom.window
+import tyrian.runtime.TyrianRuntime
 
 import scala.annotation.nowarn
 import scala.scalajs.js.Promise
@@ -110,7 +111,7 @@ trait TyrianAppF[F[_]: Async, Msg, Model]:
 
   def ready(node: Element, flags: Map[String, String]): Unit =
     run(
-      Tyrian.start[F, Model, Msg](
+      TyrianAppF.start[F, Model, Msg](
         node,
         router,
         _init(flags),
@@ -165,3 +166,35 @@ object TyrianAppF:
       case (dataAttr, attrValue) if dataAttr.startsWith("tyrianFlag") =>
         dataAttr.replaceFirst("^tyrianFlag", "") -> attrValue
     }.toMap
+
+  /** Directly starts the app. Computes the initial state of the given application, renders it on the given DOM element,
+    * and listens to user actions
+    * @param init
+    *   initial state
+    * @param update
+    *   state transition function
+    * @param view
+    *   view function
+    * @param subscriptions
+    *   subscriptions function
+    * @param node
+    *   the DOM element to mount the app to
+    * @param runner
+    *   the function that runs the program. Has a type of `F[Option[Msg]] => (Either[Throwable, Option[Msg]] => Unit) =>
+    *   Unit`, essentially: `task.unsafeRunAsync(callback)`
+    * @tparam F
+    *   The effect type to use, e.g. `IO`
+    * @tparam Model
+    *   Type of model
+    * @tparam Msg
+    *   Type of messages
+    */
+  def start[F[_]: Async, Model, Msg](
+      node: Element,
+      router: Location => Msg,
+      init: (Model, Cmd[F, Msg]),
+      update: Model => Msg => (Model, Cmd[F, Msg]),
+      view: Model => Html[Msg],
+      subscriptions: Model => Sub[F, Msg]
+  ): F[Nothing] =
+    TyrianRuntime[F, Model, Msg](router, node, init._1, init._2, update, view, subscriptions)
