+++ tyrian/src/main/scala/tyrian/cmds/FileReader.scala
