+++ tyrian/src/main/scala/tyrian/websocket/WebSocketEvent.scala
