+++ tyrian/src/main/scala/util/Functions.scala
