+++ tyrian/src/test/scala/tyrian/CmdLawsTest.scala
