+++ tyrian/src/test/scala/tyrian/CmdSubUtils.scala
