+++ tyrian/src/test/scala/tyrian/CmdTests.scala
