+++ tyrian/src/test/scala/tyrian/SubLawsTest.scala
