+++ tyrian/src/test/scala/tyrian/cmds/RandomTests.scala
